+++ mol_gnn/models/__init__.py
@@ -1,2 +1,2 @@
-from .model import MPNN
+from .mpnn import MPNN
 from .multi import MulticomponentMPNN
